pyproject.toml
[tool.poetry]
name = "sentiment analysis"
version = "0.1.0"
description = ""
authors = ["wil <psyws2@nottingham.ac.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
nltk = "^3.7"
praw = "^7.6.1"
numpy = "^1.23.5"
scipy = "^1.9.3"
transformers = "^4.24.0"
fast-ml = "^3.68"
pandas = "^1.5.1"
python-dotenv = "^0.21.0"
evaluate = "^0.3.0"
scikit-learn = "^1.1.3"
matplotlib = "^3.6.2"
fastapi = "^0.88.0"
uvicorn = "^0.20.0"
black = "^22.10.0"
flake8-black = "^0.3.5"
schedule = "^1.1.0"
torch = "^1.13.0"
torchvision = "^0.14.0"
torchaudio = "^0.13.0"
firebase-admin = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
